BUG: Fix the ``geometry_column`` issue in ``WFS`` [skip ci]
Fix an issue in ``WFS`` where the ``getfeature_bygeom`` method
fails if the requested web service does not have ``geometry_column``
attribute in its schema. This release addresses this issue by
trying to find the name from other attributes in the schema.
If it fails to find, it raises a ``ValueError``.